Fixes bridge utility commands and index calculation
Corrects discrepancies in bridge utility commands,
specifically in origin/destination network values and index
calculations.

Updates `is_claimed` to utilize the AggKit API for claim status,
bypassing contract calls for improved accuracy.

Introduces `network-id` utility command to fetch the bridge
contract's network ID.

Adjusts index calculations to align with expected global index logic,
addressing potential issues related to claim verification.

+++ cli/src/commands/bridge/mod.rs
@@ -46,7 +46,7 @@ abigen!(
         function precalculatedWrapperAddress(uint32 originNetwork, address originTokenAddress, string name, string symbol, uint8 decimals) external view returns (address)
         function getTokenWrappedAddress(uint32 originNetwork, address originTokenAddress) external view returns (address)
         function wrappedTokenToTokenInfo(address wrappedToken) external view returns (uint32, address)
-        function isClaimed(uint256 leafIndex, uint32 sourceBridgeNetwork) external view returns (bool)
+        function isClaimed(uint32 leafIndex, uint32 sourceBridgeNetwork) external view returns (bool)
         function networkID() external view returns (uint32)
     ]"#,
 );
